Move Read/WriteHandle and use pub(super) when possible
